+++ temp.py
@@ -5,13 +5,13 @@ os.environ["OPENCV_IO_ENABLE_OPENEXR"]="1"
 import numpy
 import onnxruntime
 from onnxruntime.quantization import CalibrationDataReader, quantize_static, QuantType, QuantFormat
-import loader
-import data
+import data.pano_loader as ploader
+import data.data as data
 
 import tensorflow as tf
 
 pano_path = "/data3/awong/pano/M3D_low/"
-train = loader.generate_dataframe("./splits/M3D_v1_train.yaml", pano_path)
+train = ploader.generate_dataframe("./splits/M3D_v1_train.yaml", pano_path)
 train_generator = data.DataGenerator(
             train, batch_size=32, shuffle=False, wrap=64
         )
@@ -20,11 +20,10 @@ class MobileNetDataReader(CalibrationDataReader):
     def __init__(self, model_path):
         self.enum_data = None
 
-#         # Use inference session to get input shape.
+        # Use inference session to get input shape.
         session = onnxruntime.InferenceSession(model_path, None)
         (_, _, height, width) = session.get_inputs()[0].shape
         self.input_name = session.get_inputs()[0].name
-#         self.datasize = len(train.images)
 
     def get_next(self):
         if self.enum_data is None:
